Extract route handlers into api module
