+++ packages/director/src/drivers/index.ts
@@ -1,8 +1,25 @@
-import { ExecutionDriver } from '@src/types';
+import { ExecutionDriver, ScreenshotsDriver } from '@src/types';
 import { EXECUTION_DRIVER, SCREENSHOTS_DRIVER } from '@src/config';
 
-export const getScreenshotsDriver = async (): Promise<ExecutionDriver> =>
-  import(SCREENSHOTS_DRIVER).then(module => module.driver);
+let executionDriver: ExecutionDriver | null = null;
+let storageDriver: ScreenshotsDriver | null = null;
 
-export const getExecutionDriver = async (): Promise<ExecutionDriver> =>
-  import(EXECUTION_DRIVER).then(module => module.driver);
+export const getScreenshotsDriver = async (): Promise<ScreenshotsDriver> => {
+  if (storageDriver) {
+    return storageDriver;
+  }
+  const module = await import(SCREENSHOTS_DRIVER);
+  storageDriver = module.driver;
+  await storageDriver.init();
+  return storageDriver;
+};
+
+export const getExecutionDriver = async (): Promise<ExecutionDriver> => {
+  if (executionDriver) {
+    return executionDriver;
+  }
+  const module = await import(EXECUTION_DRIVER);
+  executionDriver = module.driver;
+  await executionDriver.init();
+  return executionDriver;
+};
