+++ packages/director/src/index.ts
@@ -1,19 +1,8 @@
-import { getExecutionDriver, getScreenshotsDriver } from '@src/drivers';
 import { app } from './app';
 import { PORT } from './config';
 export * from './types';
 
 async function main() {
-  const executionDriver = await getExecutionDriver();
-  console.log(`Initializing "${executionDriver.id}" execution driver...`);
-  await executionDriver.init();
-
-  const screenshotsDriver = await getScreenshotsDriver();
-  console.log(`Initializing "${screenshotsDriver.id}" screenshots driver...`);
-  await screenshotsDriver.init();
-
-  app.set('executionDriver', executionDriver);
-  app.set('screenshotsDriver', screenshotsDriver);
   app.on('error', (error) => {
     throw error;
   });
