+++ packages/director/src/lib/allowedKeys.ts
@@ -1,0 +1,4 @@
+import { ALLOWED_KEYS } from '@src/config';
+
+export const isKeyAllowed = (recordKey: string) =>
+  ALLOWED_KEYS ? ALLOWED_KEYS.includes(recordKey) : true;
