+++ packages/director/src/types/index.ts
@@ -2,3 +2,4 @@ export * from './driver.types';
 export * from './instance.types';
 export * from './run.types';
 export * from './project.types';
+export * from './response.types';
